convert integration test into unit test
